moved non-config functions into a new lisp directory
We will now put non-config related functions (any elisp libraries or
snippets that I write) into a new ".emacs.d/lisp" directory and separate them from
the config code in the ".emacs.d/config" directory.

During the transition I decided to move and rewrite the logic that
adds all of the libraries to load-path. Now this logic is in two top
level files (load-config.el load-local-lisp.el). I needed to remove
the config-programming-languages module because it conflicted with the
new load-path logic (it was mostly useless anyways).

The man advice functions are deleted in this commit but they will be
added into the new ".emacs.d/lisp" directory soon.