deleted old themes, created new badgar theme
Themes in Emacs can be loaded on top of each other, allowing to easily
customize or override aspects of other themes, which I did not realize
at the time of creating the custom-wombat theme! Badgar adds a few
small tweaks, like setting highlight-indentation colors, making the
fringe match the background, and highlighting the current line when
using line-number-mode.